requirements.txt
# Python 3.10+ strongly recommended (official MCP PyPI package requires 3.10+).
google-adk>=1.0.0
mcp>=1.0.0; python_version>='3.10'
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
aiosqlite>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
greenlet>=3.0.0
